mv deserialize_bitmap to common-io.
