+++ src/tests/sqlsmith/src/sql_gen/dml.rs
@@ -38,7 +38,6 @@ use databend_common_ast::ast::UnmatchedClause;
 use databend_common_ast::ast::UpdateExpr;
 use databend_common_ast::ast::UpdateStmt;
 use databend_common_ast::Span;
-use databend_common_expression::types::deserialize_bitmap;
 use databend_common_expression::types::DataType;
 use databend_common_expression::Column;
 use databend_common_expression::ScalarRef;
@@ -50,6 +49,7 @@ use databend_common_io::constants::INF_BYTES_LOWER;
 use databend_common_io::constants::NAN_BYTES_LOWER;
 use databend_common_io::constants::NULL_BYTES_UPPER;
 use databend_common_io::constants::TRUE_BYTES_LOWER;
+use databend_common_io::deserialize_bitmap;
 use databend_common_sql::resolve_type_name;
 use itertools::join;
 use rand::Rng;
